requirements.txt
torch
torchvision
diffusers
transformers
accelerate
sentencepiece
tqdm
omegaconf
wandb
mosaicml-streaming
bs4
ftfy